Removed presets from "F" mode
This was done to try and fix the full-auto error on the "F" mode.
Users were unable to use the "F" mode for anything other than full-auto despite the settings.
This turned out to be the loadPreset() routine. For some reason the F/FA switch is constantly calling the loadPreset() routine.
This would reset the trigger state causing the trigger to always look freshly pulled.